@@ -1,4 +1,4 @@
-﻿// This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+// This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
 // If a copy of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // Copyright (C) PCHardwareMonitor and Contributors.
 // All Rights Reserved.
@@ -11,8 +11,8 @@ namespace PCHardwareMonitor.Hardware.Controller.Nzxt
 {
     internal class NzxtGroup : IGroup
     {
-        private readonly List<IHardware> _hardware = new List<IHardware>();
-        private readonly StringBuilder _report = new StringBuilder();
+        private readonly List<IHardware> _hardware = new();
+        private readonly StringBuilder _report = new();
 
         public NzxtGroup(ISettings settings)
         {
@@ -26,7 +26,6 @@ namespace PCHardwareMonitor.Hardware.Controller.Nzxt
                 switch (dev.ProductID)
                 {
                     case 0x2007:
-                    {
                         var device = new KrakenX3(dev, settings);
                         _report.AppendLine($"Device name: {productName}");
                         _report.AppendLine($"Firmware version: {device.FirmwareVersion}");
@@ -34,13 +33,11 @@ namespace PCHardwareMonitor.Hardware.Controller.Nzxt
                         _report.AppendLine();
                         _hardware.Add(device);
                         break;
-                    }
+
                     default:
-                    {
                         _report.AppendLine($"Unknown Hardware PID: {dev.ProductID} Name: {productName}");
                         _report.AppendLine();
                         break;
-                    }
                 }
             }
 
