@@ -18,12 +18,11 @@ namespace PCHardwareMonitor.Hardware.CPU
         private readonly Sensor[] _coreClocks;
         private readonly Sensor _coreMax;
         private readonly Sensor[] _coreTemperatures;
-        private readonly Sensor _coreVoltage;
         private readonly Sensor[] _coreVIDs;
+        private readonly Sensor _coreVoltage;
         private readonly Sensor[] _distToTjMaxTemperatures;
 
         private readonly uint[] _energyStatusMsrs = { MSR_PKG_ENERGY_STATUS, MSR_PP0_ENERGY_STATUS, MSR_PP1_ENERGY_STATUS, MSR_DRAM_ENERGY_STATUS };
-        private readonly float _energyUnitMultiplier;
         private readonly uint[] _lastEnergyConsumed;
         private readonly DateTime[] _lastEnergyTime;
 
@@ -31,8 +30,6 @@ namespace PCHardwareMonitor.Hardware.CPU
         private readonly Sensor _packageTemperature;
         private readonly Sensor[] _powerSensors;
         private readonly double _timeStampCounterMultiplier;
-
-        public float EnergyUnitsMultiplier => _energyUnitMultiplier;
 
         public IntelCpu(int processorIndex, CpuId[][] cpuId, ISettings settings) : base(processorIndex, cpuId, settings)
         {
@@ -44,174 +41,183 @@ namespace PCHardwareMonitor.Hardware.CPU
             switch (_family)
             {
                 case 0x06:
-                {
-                    switch (_model)
                     {
-                        case 0x0F: // Intel Core 2 (65nm)
-                            _microArchitecture = MicroArchitecture.Core;
-                            switch (_stepping)
-                            {
-                                case 0x06: // B2
-                                    switch (_coreCount)
+                        switch (_model)
+                        {
+                            case 0x0F: // Intel Core 2 (65nm)
+                                _microArchitecture = MicroArchitecture.Core;
+                                tjMax = _stepping switch
+                                {
+                                    // B2
+                                    0x06 => _coreCount switch
                                     {
-                                        case 2:
-                                            tjMax = Floats(80 + 10);
-                                            break;
-                                        case 4:
-                                            tjMax = Floats(90 + 10);
-                                            break;
-                                        default:
-                                            tjMax = Floats(85 + 10);
-                                            break;
-                                    }
-                                    break;
-                                case 0x0B: // G0
-                                    tjMax = Floats(90 + 10);
-                                    break;
-                                case 0x0D: // M0
-                                    tjMax = Floats(85 + 10);
-                                    break;
-                                default:
-                                    tjMax = Floats(85 + 10);
-                                    break;
-                            }
-                            break;
-                        case 0x17: // Intel Core 2 (45nm)
-                            _microArchitecture = MicroArchitecture.Core;
-                            tjMax = Floats(100);
-                            break;
-                        case 0x1C: // Intel Atom (45nm)
-                            _microArchitecture = MicroArchitecture.Atom;
-                            switch (_stepping)
-                            {
-                                case 0x02: // C0
-                                    tjMax = Floats(90);
-                                    break;
-                                case 0x0A: // A0, B0
-                                    tjMax = Floats(100);
-                                    break;
-                                default:
-                                    tjMax = Floats(90);
-                                    break;
-                            }
-                            break;
-                        case 0x1A: // Intel Core i7 LGA1366 (45nm)
-                        case 0x1E: // Intel Core i5, i7 LGA1156 (45nm)
-                        case 0x1F: // Intel Core i5, i7
-                        case 0x25: // Intel Core i3, i5, i7 LGA1156 (32nm)
-                        case 0x2C: // Intel Core i7 LGA1366 (32nm) 6 Core
-                        case 0x2E: // Intel Xeon Processor 7500 series (45nm)
-                        case 0x2F: // Intel Xeon Processor (32nm)
-                            _microArchitecture = MicroArchitecture.Nehalem;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x2A: // Intel Core i5, i7 2xxx LGA1155 (32nm)
-                        case 0x2D: // Next Generation Intel Xeon, i7 3xxx LGA2011 (32nm)
-                            _microArchitecture = MicroArchitecture.SandyBridge;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x3A: // Intel Core i5, i7 3xxx LGA1155 (22nm)
-                        case 0x3E: // Intel Core i7 4xxx LGA2011 (22nm)
-                            _microArchitecture = MicroArchitecture.IvyBridge;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x3C: // Intel Core i5, i7 4xxx LGA1150 (22nm)
-                        case 0x3F: // Intel Xeon E5-2600/1600 v3, Core i7-59xx
-                        // LGA2011-v3, Haswell-E (22nm)
-                        case 0x45: // Intel Core i5, i7 4xxxU (22nm)
-                        case 0x46:
-                            _microArchitecture = MicroArchitecture.Haswell;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x3D: // Intel Core M-5xxx (14nm)
-                        case 0x47: // Intel i5, i7 5xxx, Xeon E3-1200 v4 (14nm)
-                        case 0x4F: // Intel Xeon E5-26xx v4
-                        case 0x56: // Intel Xeon D-15xx
-                            _microArchitecture = MicroArchitecture.Broadwell;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x36: // Intel Atom S1xxx, D2xxx, N2xxx (32nm)
-                            _microArchitecture = MicroArchitecture.Atom;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x37: // Intel Atom E3xxx, Z3xxx (22nm)
-                        case 0x4A:
-                        case 0x4D: // Intel Atom C2xxx (22nm)
-                        case 0x5A:
-                        case 0x5D:
-                            _microArchitecture = MicroArchitecture.Silvermont;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x4E:
-                        case 0x5E: // Intel Core i5, i7 6xxxx LGA1151 (14nm)
-                        case 0x55: // Intel Core X i7, i9 7xxx LGA2066 (14nm)
-                            _microArchitecture = MicroArchitecture.Skylake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x4C: // Intel Airmont (Cherry Trail, Braswell)
-                            _microArchitecture = MicroArchitecture.Airmont;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x8E: // Intel Core i5, i7 7xxxx (14nm) (Kaby Lake) and 8xxxx (14nm++) (Coffee Lake)
-                        case 0x9E:
-                            _microArchitecture = MicroArchitecture.KabyLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x5C: // Goldmont (Apollo Lake)
-                        case 0x5F: // (Denverton)
-                            _microArchitecture = MicroArchitecture.Goldmont;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x7A: // Goldmont plus (Gemini Lake)
-                            _microArchitecture = MicroArchitecture.GoldmontPlus;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x66: // Intel Core i3 8xxx (10nm) (Cannon Lake)
-                            _microArchitecture = MicroArchitecture.CannonLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x7D: // Intel Core i3, i5, i7 10xxx (10nm) (Ice Lake)
-                        case 0x7E:
-                        case 0x6A: // Ice Lake server
-                        case 0x6C:
-                            _microArchitecture = MicroArchitecture.IceLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0xA5:
-                        case 0xA6: // Intel Core i3, i5, i7 10xxxU (14nm)
-                            _microArchitecture = MicroArchitecture.CometLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x86: // Tremont (10nm) (Elkhart Lake, Skyhawk Lake)
-                            _microArchitecture = MicroArchitecture.Tremont;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x8C: // Tiger Lake (10nm)
-                        case 0x8D:
-                            _microArchitecture = MicroArchitecture.TigerLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x97: // Alder Lake (7nm)
-                            _microArchitecture = MicroArchitecture.AlderLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0x9C: // Jasper Lake (10nm)
-                            _microArchitecture = MicroArchitecture.JasperLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        case 0xA7: // Intel Core i5, i6, i7 11xxx (14nm) (Rocket Lake)
-                            _microArchitecture = MicroArchitecture.RocketLake;
-                            tjMax = GetTjMaxFromMsr();
-                            break;
-                        default:
-                            _microArchitecture = MicroArchitecture.Unknown;
-                            tjMax = Floats(100);
-                            break;
+                                        2 => Floats(80 + 10),
+                                        4 => Floats(90 + 10),
+                                        _ => Floats(85 + 10),
+                                    },
+                                    // G0
+                                    0x0B => Floats(90 + 10),
+                                    // M0
+                                    0x0D => Floats(85 + 10),
+                                    _ => Floats(85 + 10),
+                                };
+                                break;
+
+                            case 0x17: // Intel Core 2 (45nm)
+                                _microArchitecture = MicroArchitecture.Core;
+                                tjMax = Floats(100);
+                                break;
+
+                            case 0x1C: // Intel Atom (45nm)
+                                _microArchitecture = MicroArchitecture.Atom;
+                                tjMax = _stepping switch
+                                {
+                                    // C0
+                                    0x02 => Floats(90),
+                                    // A0, B0
+                                    0x0A => Floats(100),
+                                    _ => Floats(90),
+                                };
+                                break;
+
+                            case 0x1A: // Intel Core i7 LGA1366 (45nm)
+                            case 0x1E: // Intel Core i5, i7 LGA1156 (45nm)
+                            case 0x1F: // Intel Core i5, i7
+                            case 0x25: // Intel Core i3, i5, i7 LGA1156 (32nm)
+                            case 0x2C: // Intel Core i7 LGA1366 (32nm) 6 Core
+                            case 0x2E: // Intel Xeon Processor 7500 series (45nm)
+                            case 0x2F: // Intel Xeon Processor (32nm)
+                                _microArchitecture = MicroArchitecture.Nehalem;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x2A: // Intel Core i5, i7 2xxx LGA1155 (32nm)
+                            case 0x2D: // Next Generation Intel Xeon, i7 3xxx LGA2011 (32nm)
+                                _microArchitecture = MicroArchitecture.SandyBridge;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x3A: // Intel Core i5, i7 3xxx LGA1155 (22nm)
+                            case 0x3E: // Intel Core i7 4xxx LGA2011 (22nm)
+                                _microArchitecture = MicroArchitecture.IvyBridge;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x3C: // Intel Core i5, i7 4xxx LGA1150 (22nm)
+                            case 0x3F: // Intel Xeon E5-2600/1600 v3, Core i7-59xx
+                            // LGA2011-v3, Haswell-E (22nm)
+                            case 0x45: // Intel Core i5, i7 4xxxU (22nm)
+                            case 0x46:
+                                _microArchitecture = MicroArchitecture.Haswell;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x3D: // Intel Core M-5xxx (14nm)
+                            case 0x47: // Intel i5, i7 5xxx, Xeon E3-1200 v4 (14nm)
+                            case 0x4F: // Intel Xeon E5-26xx v4
+                            case 0x56: // Intel Xeon D-15xx
+                                _microArchitecture = MicroArchitecture.Broadwell;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x36: // Intel Atom S1xxx, D2xxx, N2xxx (32nm)
+                                _microArchitecture = MicroArchitecture.Atom;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x37: // Intel Atom E3xxx, Z3xxx (22nm)
+                            case 0x4A:
+                            case 0x4D: // Intel Atom C2xxx (22nm)
+                            case 0x5A:
+                            case 0x5D:
+                                _microArchitecture = MicroArchitecture.Silvermont;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x4E:
+                            case 0x5E: // Intel Core i5, i7 6xxxx LGA1151 (14nm)
+                            case 0x55: // Intel Core X i7, i9 7xxx LGA2066 (14nm)
+                                _microArchitecture = MicroArchitecture.Skylake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x4C: // Intel Airmont (Cherry Trail, Braswell)
+                                _microArchitecture = MicroArchitecture.Airmont;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x8E: // Intel Core i5, i7 7xxxx (14nm) (Kaby Lake) and 8xxxx (14nm++) (Coffee Lake)
+                            case 0x9E:
+                                _microArchitecture = MicroArchitecture.KabyLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x5C: // Goldmont (Apollo Lake)
+                            case 0x5F: // (Denverton)
+                                _microArchitecture = MicroArchitecture.Goldmont;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x7A: // Goldmont plus (Gemini Lake)
+                                _microArchitecture = MicroArchitecture.GoldmontPlus;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x66: // Intel Core i3 8xxx (10nm) (Cannon Lake)
+                                _microArchitecture = MicroArchitecture.CannonLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x7D: // Intel Core i3, i5, i7 10xxx (10nm) (Ice Lake)
+                            case 0x7E:
+                            case 0x6A: // Ice Lake server
+                            case 0x6C:
+                                _microArchitecture = MicroArchitecture.IceLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0xA5:
+                            case 0xA6: // Intel Core i3, i5, i7 10xxxU (14nm)
+                                _microArchitecture = MicroArchitecture.CometLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x86: // Tremont (10nm) (Elkhart Lake, Skyhawk Lake)
+                                _microArchitecture = MicroArchitecture.Tremont;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x8C: // Tiger Lake (10nm)
+                            case 0x8D:
+                                _microArchitecture = MicroArchitecture.TigerLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x97: // Alder Lake-S (Intel 7/10nm)
+                            case 0x9A: // Alder Lake-H (Intel 7/10nm)
+                                _microArchitecture = MicroArchitecture.AlderLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0x9C: // Jasper Lake (10nm)
+                                _microArchitecture = MicroArchitecture.JasperLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            case 0xA7: // Intel Core i5, i6, i7 11xxx (14nm) (Rocket Lake)
+                                _microArchitecture = MicroArchitecture.RocketLake;
+                                tjMax = GetTjMaxFromMsr();
+                                break;
+
+                            default:
+                                _microArchitecture = MicroArchitecture.Unknown;
+                                tjMax = Floats(100);
+                                break;
+                        }
                     }
-                }
-                break;
+
+                    break;
                 case 0x0F:
-                {
                     switch (_model)
                     {
                         case 0x00: // Pentium 4 (180nm)
@@ -223,32 +229,30 @@ namespace PCHardwareMonitor.Hardware.CPU
                             _microArchitecture = MicroArchitecture.NetBurst;
                             tjMax = Floats(100);
                             break;
+
                         default:
                             _microArchitecture = MicroArchitecture.Unknown;
                             tjMax = Floats(100);
                             break;
                     }
-                }
-                break;
+
+                    break;
                 default:
                     _microArchitecture = MicroArchitecture.Unknown;
                     tjMax = Floats(100);
-                break;
+                    break;
             }
+
             // set timeStampCounterMultiplier
             switch (_microArchitecture)
             {
                 case MicroArchitecture.Atom:
                 case MicroArchitecture.Core:
                 case MicroArchitecture.NetBurst:
-                {
                     if (Ring0.ReadMsr(IA32_PERF_STATUS, out uint _, out edx))
-                    {
-                        _timeStampCounterMultiplier = ((edx >> 8) & 0x1f) + 0.5 * ((edx >> 14) & 1);
-                    }
+                        _timeStampCounterMultiplier = ((edx >> 8) & 0x1f) + (0.5 * ((edx >> 14) & 1));
 
                     break;
-                }
                 case MicroArchitecture.Airmont:
                 case MicroArchitecture.AlderLake:
                 case MicroArchitecture.Broadwell:
@@ -268,13 +272,10 @@ namespace PCHardwareMonitor.Hardware.CPU
                 case MicroArchitecture.Skylake:
                 case MicroArchitecture.TigerLake:
                 case MicroArchitecture.Tremont:
-                {
                     if (Ring0.ReadMsr(MSR_PLATFORM_INFO, out eax, out uint _))
-                    {
                         _timeStampCounterMultiplier = (eax >> 8) & 0xff;
-                    }
-                }
-                break;
+
+                    break;
                 default:
                     _timeStampCounterMultiplier = 0;
                     break;
@@ -304,7 +305,7 @@ namespace PCHardwareMonitor.Hardware.CPU
                 }
             }
             else
-                _coreTemperatures = new Sensor[0];
+                _coreTemperatures = Array.Empty<Sensor>();
 
             // check if processor supports a digital thermal sensor at package level
             if (cpuId[0][0].Data.GetLength(0) > 6 && (cpuId[0][0].Data[6, 0] & 0x40) != 0 && _microArchitecture != MicroArchitecture.Unknown)
@@ -336,7 +337,7 @@ namespace PCHardwareMonitor.Hardware.CPU
                 }
             }
             else
-                _distToTjMaxTemperatures = new Sensor[0];
+                _distToTjMaxTemperatures = Array.Empty<Sensor>();
 
             //core temp avg and max value
             //is only available when the cpu has more than 1 core
@@ -364,42 +365,39 @@ namespace PCHardwareMonitor.Hardware.CPU
                     ActivateSensor(_coreClocks[i]);
             }
 
-            if (_microArchitecture == MicroArchitecture.Airmont ||
-                _microArchitecture == MicroArchitecture.AlderLake ||
-                _microArchitecture == MicroArchitecture.Broadwell ||
-                _microArchitecture == MicroArchitecture.CannonLake ||
-                _microArchitecture == MicroArchitecture.CometLake ||
-                _microArchitecture == MicroArchitecture.Goldmont ||
-                _microArchitecture == MicroArchitecture.GoldmontPlus ||
-                _microArchitecture == MicroArchitecture.Haswell ||
-                _microArchitecture == MicroArchitecture.IceLake ||
-                _microArchitecture == MicroArchitecture.IvyBridge ||
-                _microArchitecture == MicroArchitecture.JasperLake ||
-                _microArchitecture == MicroArchitecture.KabyLake ||
-                _microArchitecture == MicroArchitecture.RocketLake ||
-                _microArchitecture == MicroArchitecture.SandyBridge ||
-                _microArchitecture == MicroArchitecture.Silvermont ||
-                _microArchitecture == MicroArchitecture.Skylake ||
-                _microArchitecture == MicroArchitecture.TigerLake ||
-                _microArchitecture == MicroArchitecture.Tremont)
+            if (_microArchitecture is MicroArchitecture.Airmont or
+                MicroArchitecture.AlderLake or
+                MicroArchitecture.Broadwell or
+                MicroArchitecture.CannonLake or
+                MicroArchitecture.CometLake or
+                MicroArchitecture.Goldmont or
+                MicroArchitecture.GoldmontPlus or
+                MicroArchitecture.Haswell or
+                MicroArchitecture.IceLake or
+                MicroArchitecture.IvyBridge or
+                MicroArchitecture.JasperLake or
+                MicroArchitecture.KabyLake or
+                MicroArchitecture.RocketLake or
+                MicroArchitecture.SandyBridge or
+                MicroArchitecture.Silvermont or
+                MicroArchitecture.Skylake or
+                MicroArchitecture.TigerLake or
+                MicroArchitecture.Tremont)
             {
                 _powerSensors = new Sensor[_energyStatusMsrs.Length];
                 _lastEnergyTime = new DateTime[_energyStatusMsrs.Length];
                 _lastEnergyConsumed = new uint[_energyStatusMsrs.Length];
 
                 if (Ring0.ReadMsr(MSR_RAPL_POWER_UNIT, out eax, out uint _))
-                    switch (_microArchitecture)
+                {
+                    EnergyUnitsMultiplier = _microArchitecture switch
                     {
-                        case MicroArchitecture.Silvermont:
-                        case MicroArchitecture.Airmont:
-                            _energyUnitMultiplier = 1.0e-6f * (1 << (int)((eax >> 8) & 0x1F));
-                            break;
-                        default:
-                            _energyUnitMultiplier = 1.0f / (1 << (int)((eax >> 8) & 0x1F));
-                            break;
-                    }
+                        MicroArchitecture.Silvermont or MicroArchitecture.Airmont => 1.0e-6f * (1 << (int)((eax >> 8) & 0x1F)),
+                        _ => 1.0f / (1 << (int)((eax >> 8) & 0x1F)),
+                    };
+                }
 
-                if (_energyUnitMultiplier != 0)
+                if (EnergyUnitsMultiplier != 0)
                 {
                     string[] powerSensorLabels = { "CPU Package", "CPU Cores", "CPU Graphics", "CPU Memory" };
 
@@ -440,6 +438,8 @@ namespace PCHardwareMonitor.Hardware.CPU
 
             Update();
         }
+
+        public float EnergyUnitsMultiplier { get; }
 
         private float[] Floats(float f)
         {
@@ -483,7 +483,7 @@ namespace PCHardwareMonitor.Hardware.CPU
 
         public override string GetReport()
         {
-            StringBuilder r = new StringBuilder();
+            StringBuilder r = new();
             r.Append(base.GetReport());
             r.Append("MicroArchitecture: ");
             r.AppendLine(_microArchitecture.ToString());
@@ -511,7 +511,7 @@ namespace PCHardwareMonitor.Hardware.CPU
                     float deltaT = (eax & 0x007F0000) >> 16;
                     float tjMax = _coreTemperatures[i].Parameters[0].Value;
                     float tSlope = _coreTemperatures[i].Parameters[1].Value;
-                    _coreTemperatures[i].Value = tjMax - tSlope * deltaT;
+                    _coreTemperatures[i].Value = tjMax - (tSlope * deltaT);
 
                     coreAvg += (float)_coreTemperatures[i].Value;
                     if (coreMax < _coreTemperatures[i].Value)
@@ -543,7 +543,7 @@ namespace PCHardwareMonitor.Hardware.CPU
                     float deltaT = (eax & 0x007F0000) >> 16;
                     float tjMax = _packageTemperature.Parameters[0].Value;
                     float tSlope = _packageTemperature.Parameters[1].Value;
-                    _packageTemperature.Value = tjMax - tSlope * deltaT;
+                    _packageTemperature.Value = tjMax - (tSlope * deltaT);
                 }
                 else
                 {
@@ -563,11 +563,8 @@ namespace PCHardwareMonitor.Hardware.CPU
                         switch (_microArchitecture)
                         {
                             case MicroArchitecture.Nehalem:
-                            {
-                                uint multiplier = eax & 0xff;
-                                _coreClocks[i].Value = (float)(multiplier * newBusClock);
+                                _coreClocks[i].Value = (float)((eax & 0xff) * newBusClock);
                                 break;
-                            }
                             case MicroArchitecture.Airmont:
                             case MicroArchitecture.AlderLake:
                             case MicroArchitecture.Broadwell:
@@ -586,17 +583,11 @@ namespace PCHardwareMonitor.Hardware.CPU
                             case MicroArchitecture.Skylake:
                             case MicroArchitecture.TigerLake:
                             case MicroArchitecture.Tremont:
-                            {
-                                uint multiplier = (eax >> 8) & 0xff;
-                                _coreClocks[i].Value = (float)(multiplier * newBusClock);
+                                _coreClocks[i].Value = (float)(((eax >> 8) & 0xff) * newBusClock);
                                 break;
-                            }
                             default:
-                            {
-                                double multiplier = ((eax >> 8) & 0x1f) + 0.5 * ((eax >> 14) & 1);
-                                _coreClocks[i].Value = (float)(multiplier * newBusClock);
+                                _coreClocks[i].Value = (float)((((eax >> 8) & 0x1f) + (0.5 * ((eax >> 14) & 1))) * newBusClock);
                                 break;
-                            }
                         }
                     }
                     else
@@ -623,15 +614,13 @@ namespace PCHardwareMonitor.Hardware.CPU
                     if (!Ring0.ReadMsr(_energyStatusMsrs[sensor.Index], out eax, out uint _))
                         continue;
 
-
                     DateTime time = DateTime.UtcNow;
                     uint energyConsumed = eax;
                     float deltaTime = (float)(time - _lastEnergyTime[sensor.Index]).TotalSeconds;
                     if (deltaTime < 0.01)
                         continue;
 
-
-                    sensor.Value = _energyUnitMultiplier * unchecked(energyConsumed - _lastEnergyConsumed[sensor.Index]) / deltaTime;
+                    sensor.Value = EnergyUnitsMultiplier * unchecked(energyConsumed - _lastEnergyConsumed[sensor.Index]) / deltaTime;
                     _lastEnergyTime[sensor.Index] = time;
                     _lastEnergyConsumed[sensor.Index] = energyConsumed;
                 }
