@@ -359,8 +359,7 @@ namespace PCHardwareMonitor.Hardware
                         w.Write(report);
                     }
 
-                    IEnumerable<IHardware> hardwareArray = group.Hardware;
-                    foreach (IHardware hardware in hardwareArray)
+                    foreach (IHardware hardware in (IEnumerable<IHardware>)group.Hardware)
                         ReportHardware(hardware, w);
                 }
 
