@@ -77,7 +77,7 @@ namespace PCHardwareMonitor.Hardware.Gpu
                         _ => "GPU"
                     };
 
-                    _temperatures[i] = new Sensor(name, i, SensorType.Temperature, this, new ParameterDescription[0], settings);
+                    _temperatures[i] = new Sensor(name, i, SensorType.Temperature, this, Array.Empty<ParameterDescription>(), settings);
                     ActivateSensor(_temperatures[i]);
                 }
             }
@@ -990,7 +990,7 @@ namespace PCHardwareMonitor.Hardware.Gpu
 
             NvApi.NvThermalSettings settings = new()
             {
-                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvThermalSettings>(1), 
+                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvThermalSettings>(1),
                 Count = NvApi.MAX_THERMAL_SENSORS_PER_GPU,
             };
 
@@ -1008,7 +1008,7 @@ namespace PCHardwareMonitor.Hardware.Gpu
 
             var thermalSensors = new NvApi.NvThermalSensors
             {
-                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvThermalSensors>(2), 
+                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvThermalSensors>(2),
                 Mask = mask
             };
 
@@ -1026,7 +1026,7 @@ namespace PCHardwareMonitor.Hardware.Gpu
 
             var coolers = new NvApi.NvFanCoolersStatus
             {
-                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvFanCoolersStatus>(1), 
+                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvFanCoolersStatus>(1),
                 Items = new NvApi.NvFanCoolersStatusItem[NvApi.MAX_FAN_COOLERS_STATUS_ITEMS]
             };
 
@@ -1061,7 +1061,7 @@ namespace PCHardwareMonitor.Hardware.Gpu
 
             NvApi.NvCoolerSettings settings = new()
             {
-                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvCoolerSettings>(2), 
+                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvCoolerSettings>(2),
                 Cooler = new NvApi.NvCooler[NvApi.MAX_COOLERS_PER_GPU]
             };
 
@@ -1079,7 +1079,7 @@ namespace PCHardwareMonitor.Hardware.Gpu
 
             NvApi.NvDynamicPStatesInfo pStatesInfo = new()
             {
-                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvDynamicPStatesInfo>(1), 
+                Version = (uint)NvApi.MAKE_NVAPI_VERSION<NvApi.NvDynamicPStatesInfo>(1),
                 Utilizations = new NvApi.NvDynamicPState[NvApi.MAX_GPU_UTILIZATIONS]
             };
 
