@@ -1,4 +1,4 @@
-﻿// This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+// This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
 // If a copy of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // Copyright (C) PCHardwareMonitor and Contributors.
 // All Rights Reserved.
@@ -104,10 +104,9 @@ namespace PCHardwareMonitor.Hardware.Storage
 
         protected void Dispose(bool disposing)
         {
-            if (disposing)
+            if (disposing && _handle is { IsClosed: false })
             {
-                if (_handle != null && !_handle.IsClosed)
-                    _handle.Close();
+                _handle.Close();
             }
         }
 
