@@ -43,8 +43,7 @@ namespace PCHardwareMonitor.Hardware.Storage
                 //map NVME_IDENTIFY_CONTROLLER_DATA to nptwb.Buffer
                 IntPtr offset = Marshal.OffsetOf<Kernel32.STORAGE_QUERY_BUFFER>(nameof(Kernel32.STORAGE_QUERY_BUFFER.Buffer));
                 var newPtr = IntPtr.Add(buffer, offset.ToInt32());
-                Kernel32.NVME_IDENTIFY_CONTROLLER_DATA item = Marshal.PtrToStructure<Kernel32.NVME_IDENTIFY_CONTROLLER_DATA>(newPtr);
-                data = item;
+                data = Marshal.PtrToStructure<Kernel32.NVME_IDENTIFY_CONTROLLER_DATA>(newPtr);
                 Marshal.FreeHGlobal(buffer);
                 result = true;
             }
@@ -81,8 +80,7 @@ namespace PCHardwareMonitor.Hardware.Storage
                 //map NVME_HEALTH_INFO_LOG to nptwb.Buffer
                 IntPtr offset = Marshal.OffsetOf<Kernel32.STORAGE_QUERY_BUFFER>(nameof(Kernel32.STORAGE_QUERY_BUFFER.Buffer));
                 var newPtr = IntPtr.Add(buffer, offset.ToInt32());
-                Kernel32.NVME_HEALTH_INFO_LOG item = Marshal.PtrToStructure<Kernel32.NVME_HEALTH_INFO_LOG>(newPtr);
-                data = item;
+                data = Marshal.PtrToStructure<Kernel32.NVME_HEALTH_INFO_LOG>(newPtr);
                 Marshal.FreeHGlobal(buffer);
                 result = true;
             }
